Improve code quality

+++ src/com/github/antoinejt/jasc/calculator/CalculatorEngine.java
@@ -85,7 +85,7 @@ public final class CalculatorEngine {
                 case MULTIPLICATION: calc = operand[1] * operand[0]; break;
                 case DIVISION: calc = operand[1] / operand[0]; break;
                 case MODULO: calc = operand[1] % operand[0]; break;
-                case POWER: calc = Double.valueOf(Math.pow(operand[1], operand[0])).floatValue(); break;
+                case POWER: calc = (float) Math.pow(operand[1], operand[0]); break;
                 default: throw new CalculatorException("The provided operation is not handled!");
             }
             // I can't imagine how this can happened
